pyproject.toml
[project]
name = "dpvo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "evo>=1.33.0",
    "kornia>=0.8.2",
    "numba>=0.62.1",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "plyfile>=1.1.3",
    "pypose>=0.7.3",
    "rerun-sdk>=0.26.2",
    "tensorboard>=2.20.0",
    "torch==2.3.1",
    "torch-scatter>=2.1.2",
    "torchvision>=0.18.1",
    "tqdm>=4.67.1",
    "yacs>=0.1.8",
]

[build-system]
requires = ["setuptools>=61", "wheel", "torch==2.3.1"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
torch-scatter = { index = "pyg-cu121-pt23" }

[[tool.uv.index]]
name = "pyg-cu121-pt23"
url = "https://data.pyg.org/whl/torch-2.3.1+cu121.html"
format = "flat"
explicit = true

